download the given content as a file.
