+++ packages/commonwealth/client/scripts/utils/downloadDataAsFile.ts
@@ -1,0 +1,19 @@
+/**
+ * Save the content locally and use the `filename` as the suggested name for the
+ * file.
+ */
+export function downloadDataAsFile(content: string, filename: string) {
+  const blob = new Blob([content], { type: 'text/markdown' });
+
+  const link = document.createElement('a');
+  link.href = URL.createObjectURL(blob);
+  link.download = filename;
+
+  document.body.appendChild(link);
+
+  link.click();
+
+  document.body.removeChild(link);
+
+  URL.revokeObjectURL(link.href);
+}
